ref: disassociate envelope creation with transaction type
